query-input/requirements.txt
openpyxl
psycopg2-binary
python-dotenv
